Move bpf loader upgradeable ix processing to new module
