+++ programs/bpf_loader/src/upgradeable/mod.rs
@@ -1,3 +1,4 @@
+pub mod processor;
 pub mod with_jit;
 
 solana_sdk::declare_builtin!(
