+++ programs/bpf_loader/src/upgradeable/processor.rs
@@ -1,0 +1,2729 @@
+use {
+    crate::{create_executor, write_program_data},
+    solana_program_runtime::{
+        ic_logger_msg, invoke_context::InvokeContext, log_collector::LogCollector,
+        sysvar_cache::get_sysvar_with_account_check,
+    },
+    solana_sdk::{
+        bpf_loader_upgradeable::UpgradeableLoaderState,
+        feature_set::{disable_deploy_of_alloc_free_syscall, reduce_required_deploy_balance},
+        instruction::{AccountMeta, InstructionError},
+        loader_upgradeable_instruction::UpgradeableLoaderInstruction,
+        program_utils::limited_deserialize,
+        pubkey::Pubkey,
+        system_instruction::{self, MAX_PERMITTED_DATA_LENGTH},
+        transaction_context::{InstructionContext, TransactionContext},
+    },
+    std::{cell::RefCell, rc::Rc},
+};
+
+pub(crate) fn process_instruction(
+    first_instruction_account: usize,
+    invoke_context: &mut InvokeContext,
+    use_jit: bool,
+) -> Result<(), InstructionError> {
+    let log_collector = invoke_context.get_log_collector();
+    let transaction_context = &invoke_context.transaction_context;
+    let instruction_context = transaction_context.get_current_instruction_context()?;
+    let instruction_data = instruction_context.get_instruction_data();
+    let program_id = instruction_context.get_program_key(transaction_context)?;
+
+    match limited_deserialize(instruction_data)? {
+        UpgradeableLoaderInstruction::InitializeBuffer => {
+            instruction_context.check_number_of_instruction_accounts(2)?;
+            let mut buffer =
+                instruction_context.try_borrow_instruction_account(transaction_context, 0)?;
+
+            if UpgradeableLoaderState::Uninitialized != buffer.get_state()? {
+                ic_logger_msg!(log_collector, "Buffer account already initialized");
+                return Err(InstructionError::AccountAlreadyInitialized);
+            }
+
+            let authority_key = Some(
+                *transaction_context.get_key_of_account_at_index(
+                    instruction_context
+                        .get_index_in_transaction(first_instruction_account.saturating_add(1))?,
+                )?,
+            );
+
+            buffer.set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: authority_key,
+            })?;
+        }
+        UpgradeableLoaderInstruction::Write { offset, bytes } => {
+            instruction_context.check_number_of_instruction_accounts(2)?;
+            let buffer =
+                instruction_context.try_borrow_instruction_account(transaction_context, 0)?;
+
+            if let UpgradeableLoaderState::Buffer { authority_address } = buffer.get_state()? {
+                if authority_address.is_none() {
+                    ic_logger_msg!(log_collector, "Buffer is immutable");
+                    return Err(InstructionError::Immutable); // TODO better error code
+                }
+                let authority_key =
+                    Some(*transaction_context.get_key_of_account_at_index(
+                        instruction_context.get_index_in_transaction(
+                            first_instruction_account.saturating_add(1),
+                        )?,
+                    )?);
+                if authority_address != authority_key {
+                    ic_logger_msg!(log_collector, "Incorrect buffer authority provided");
+                    return Err(InstructionError::IncorrectAuthority);
+                }
+                if !instruction_context.is_signer(first_instruction_account.saturating_add(1))? {
+                    ic_logger_msg!(log_collector, "Buffer authority did not sign");
+                    return Err(InstructionError::MissingRequiredSignature);
+                }
+            } else {
+                ic_logger_msg!(log_collector, "Invalid Buffer account");
+                return Err(InstructionError::InvalidAccountData);
+            }
+            drop(buffer);
+            write_program_data(
+                first_instruction_account,
+                UpgradeableLoaderState::size_of_buffer_metadata().saturating_add(offset as usize),
+                &bytes,
+                invoke_context,
+            )?;
+        }
+        UpgradeableLoaderInstruction::DeployWithMaxDataLen { max_data_len } => {
+            instruction_context.check_number_of_instruction_accounts(4)?;
+            let payer_key = *transaction_context.get_key_of_account_at_index(
+                instruction_context.get_index_in_transaction(first_instruction_account)?,
+            )?;
+            let programdata_key = *transaction_context.get_key_of_account_at_index(
+                instruction_context
+                    .get_index_in_transaction(first_instruction_account.saturating_add(1))?,
+            )?;
+            let rent = get_sysvar_with_account_check::rent(invoke_context, instruction_context, 4)?;
+            let clock =
+                get_sysvar_with_account_check::clock(invoke_context, instruction_context, 5)?;
+            instruction_context.check_number_of_instruction_accounts(8)?;
+            let authority_key = Some(
+                *transaction_context.get_key_of_account_at_index(
+                    instruction_context
+                        .get_index_in_transaction(first_instruction_account.saturating_add(7))?,
+                )?,
+            );
+
+            // Verify Program account
+
+            let program =
+                instruction_context.try_borrow_instruction_account(transaction_context, 2)?;
+            if UpgradeableLoaderState::Uninitialized != program.get_state()? {
+                ic_logger_msg!(log_collector, "Program account already initialized");
+                return Err(InstructionError::AccountAlreadyInitialized);
+            }
+            if program.get_data().len() < UpgradeableLoaderState::size_of_program() {
+                ic_logger_msg!(log_collector, "Program account too small");
+                return Err(InstructionError::AccountDataTooSmall);
+            }
+            if program.get_lamports() < rent.minimum_balance(program.get_data().len()) {
+                ic_logger_msg!(log_collector, "Program account not rent-exempt");
+                return Err(InstructionError::ExecutableAccountNotRentExempt);
+            }
+            let new_program_id = *program.get_key();
+            drop(program);
+
+            // Verify Buffer account
+
+            let buffer =
+                instruction_context.try_borrow_instruction_account(transaction_context, 3)?;
+            if let UpgradeableLoaderState::Buffer { authority_address } = buffer.get_state()? {
+                if authority_address != authority_key {
+                    ic_logger_msg!(log_collector, "Buffer and upgrade authority don't match");
+                    return Err(InstructionError::IncorrectAuthority);
+                }
+                if !instruction_context.is_signer(first_instruction_account.saturating_add(7))? {
+                    ic_logger_msg!(log_collector, "Upgrade authority did not sign");
+                    return Err(InstructionError::MissingRequiredSignature);
+                }
+            } else {
+                ic_logger_msg!(log_collector, "Invalid Buffer account");
+                return Err(InstructionError::InvalidArgument);
+            }
+            let buffer_key = *buffer.get_key();
+            let buffer_lamports = buffer.get_lamports();
+            let buffer_data_offset = UpgradeableLoaderState::size_of_buffer_metadata();
+            let buffer_data_len = buffer.get_data().len().saturating_sub(buffer_data_offset);
+            let programdata_data_offset = UpgradeableLoaderState::size_of_programdata_metadata();
+            let programdata_len = UpgradeableLoaderState::size_of_programdata(max_data_len);
+            if buffer.get_data().len() < UpgradeableLoaderState::size_of_buffer_metadata()
+                || buffer_data_len == 0
+            {
+                ic_logger_msg!(log_collector, "Buffer account too small");
+                return Err(InstructionError::InvalidAccountData);
+            }
+            drop(buffer);
+            if max_data_len < buffer_data_len {
+                ic_logger_msg!(
+                    log_collector,
+                    "Max data length is too small to hold Buffer data"
+                );
+                return Err(InstructionError::AccountDataTooSmall);
+            }
+            if programdata_len > MAX_PERMITTED_DATA_LENGTH as usize {
+                ic_logger_msg!(log_collector, "Max data length is too large");
+                return Err(InstructionError::InvalidArgument);
+            }
+
+            // Create ProgramData account
+
+            let (derived_address, bump_seed) =
+                Pubkey::find_program_address(&[new_program_id.as_ref()], program_id);
+            if derived_address != programdata_key {
+                ic_logger_msg!(log_collector, "ProgramData address is not derived");
+                return Err(InstructionError::InvalidArgument);
+            }
+
+            let predrain_buffer = invoke_context
+                .feature_set
+                .is_active(&reduce_required_deploy_balance::id());
+            if predrain_buffer {
+                // Drain the Buffer account to payer before paying for programdata account
+                let mut payer =
+                    instruction_context.try_borrow_instruction_account(transaction_context, 0)?;
+                payer.checked_add_lamports(buffer_lamports)?;
+                drop(payer);
+                let mut buffer =
+                    instruction_context.try_borrow_instruction_account(transaction_context, 3)?;
+                buffer.set_lamports(0)?;
+            }
+
+            let mut instruction = system_instruction::create_account(
+                &payer_key,
+                &programdata_key,
+                1.max(rent.minimum_balance(programdata_len)),
+                programdata_len as u64,
+                program_id,
+            );
+
+            // pass an extra account to avoid the overly strict UnbalancedInstruction error
+            instruction
+                .accounts
+                .push(AccountMeta::new(buffer_key, false));
+
+            let transaction_context = &invoke_context.transaction_context;
+            let instruction_context = transaction_context.get_current_instruction_context()?;
+            let caller_program_id = instruction_context.get_program_key(transaction_context)?;
+            let signers = [&[new_program_id.as_ref(), &[bump_seed]]]
+                .iter()
+                .map(|seeds| Pubkey::create_program_address(*seeds, caller_program_id))
+                .collect::<Result<Vec<Pubkey>, solana_sdk::pubkey::PubkeyError>>()?;
+            invoke_context.native_invoke(instruction, signers.as_slice())?;
+
+            // Load and verify the program bits
+            let executor = create_executor(
+                first_instruction_account.saturating_add(3),
+                buffer_data_offset,
+                invoke_context,
+                use_jit,
+                true,
+                invoke_context
+                    .feature_set
+                    .is_active(&disable_deploy_of_alloc_free_syscall::id()),
+            )?;
+            invoke_context.update_executor(&new_program_id, executor);
+
+            let transaction_context = &invoke_context.transaction_context;
+            let instruction_context = transaction_context.get_current_instruction_context()?;
+
+            // Update the ProgramData account and record the program bits
+            {
+                let mut programdata =
+                    instruction_context.try_borrow_instruction_account(transaction_context, 1)?;
+                programdata.set_state(&UpgradeableLoaderState::ProgramData {
+                    slot: clock.slot,
+                    upgrade_authority_address: authority_key,
+                })?;
+                let dst_slice = programdata
+                    .get_data_mut()?
+                    .get_mut(
+                        programdata_data_offset
+                            ..programdata_data_offset.saturating_add(buffer_data_len),
+                    )
+                    .ok_or(InstructionError::AccountDataTooSmall)?;
+                let buffer =
+                    instruction_context.try_borrow_instruction_account(transaction_context, 3)?;
+                let src_slice = buffer
+                    .get_data()
+                    .get(buffer_data_offset..)
+                    .ok_or(InstructionError::AccountDataTooSmall)?;
+                dst_slice.copy_from_slice(src_slice);
+            }
+
+            // Update the Program account
+            let mut program =
+                instruction_context.try_borrow_instruction_account(transaction_context, 2)?;
+            program.set_state(&UpgradeableLoaderState::Program {
+                programdata_address: programdata_key,
+            })?;
+            program.set_executable(true)?;
+            drop(program);
+
+            if !predrain_buffer {
+                // Drain the Buffer account back to the payer
+                let mut payer =
+                    instruction_context.try_borrow_instruction_account(transaction_context, 0)?;
+                payer.checked_add_lamports(buffer_lamports)?;
+                drop(payer);
+                let mut buffer =
+                    instruction_context.try_borrow_instruction_account(transaction_context, 3)?;
+                buffer.set_lamports(0)?;
+            }
+
+            ic_logger_msg!(log_collector, "Deployed program {:?}", new_program_id);
+        }
+        UpgradeableLoaderInstruction::Upgrade => {
+            instruction_context.check_number_of_instruction_accounts(3)?;
+            let programdata_key = *transaction_context.get_key_of_account_at_index(
+                instruction_context.get_index_in_transaction(first_instruction_account)?,
+            )?;
+            let rent = get_sysvar_with_account_check::rent(invoke_context, instruction_context, 4)?;
+            let clock =
+                get_sysvar_with_account_check::clock(invoke_context, instruction_context, 5)?;
+            instruction_context.check_number_of_instruction_accounts(7)?;
+            let authority_key = Some(
+                *transaction_context.get_key_of_account_at_index(
+                    instruction_context
+                        .get_index_in_transaction(first_instruction_account.saturating_add(6))?,
+                )?,
+            );
+
+            // Verify Program account
+
+            let program =
+                instruction_context.try_borrow_instruction_account(transaction_context, 1)?;
+            if !program.is_executable() {
+                ic_logger_msg!(log_collector, "Program account not executable");
+                return Err(InstructionError::AccountNotExecutable);
+            }
+            if !program.is_writable() {
+                ic_logger_msg!(log_collector, "Program account not writeable");
+                return Err(InstructionError::InvalidArgument);
+            }
+            if program.get_owner() != program_id {
+                ic_logger_msg!(log_collector, "Program account not owned by loader");
+                return Err(InstructionError::IncorrectProgramId);
+            }
+            if let UpgradeableLoaderState::Program {
+                programdata_address,
+            } = program.get_state()?
+            {
+                if programdata_address != programdata_key {
+                    ic_logger_msg!(log_collector, "Program and ProgramData account mismatch");
+                    return Err(InstructionError::InvalidArgument);
+                }
+            } else {
+                ic_logger_msg!(log_collector, "Invalid Program account");
+                return Err(InstructionError::InvalidAccountData);
+            }
+            let new_program_id = *program.get_key();
+            drop(program);
+
+            // Verify Buffer account
+
+            let buffer =
+                instruction_context.try_borrow_instruction_account(transaction_context, 2)?;
+            if let UpgradeableLoaderState::Buffer { authority_address } = buffer.get_state()? {
+                if authority_address != authority_key {
+                    ic_logger_msg!(log_collector, "Buffer and upgrade authority don't match");
+                    return Err(InstructionError::IncorrectAuthority);
+                }
+                if !instruction_context.is_signer(first_instruction_account.saturating_add(6))? {
+                    ic_logger_msg!(log_collector, "Upgrade authority did not sign");
+                    return Err(InstructionError::MissingRequiredSignature);
+                }
+            } else {
+                ic_logger_msg!(log_collector, "Invalid Buffer account");
+                return Err(InstructionError::InvalidArgument);
+            }
+            let buffer_lamports = buffer.get_lamports();
+            let buffer_data_offset = UpgradeableLoaderState::size_of_buffer_metadata();
+            let buffer_data_len = buffer.get_data().len().saturating_sub(buffer_data_offset);
+            if buffer.get_data().len() < UpgradeableLoaderState::size_of_buffer_metadata()
+                || buffer_data_len == 0
+            {
+                ic_logger_msg!(log_collector, "Buffer account too small");
+                return Err(InstructionError::InvalidAccountData);
+            }
+            drop(buffer);
+
+            // Verify ProgramData account
+
+            let programdata =
+                instruction_context.try_borrow_instruction_account(transaction_context, 0)?;
+            let programdata_data_offset = UpgradeableLoaderState::size_of_programdata_metadata();
+            let programdata_balance_required =
+                1.max(rent.minimum_balance(programdata.get_data().len()));
+            if programdata.get_data().len()
+                < UpgradeableLoaderState::size_of_programdata(buffer_data_len)
+            {
+                ic_logger_msg!(log_collector, "ProgramData account not large enough");
+                return Err(InstructionError::AccountDataTooSmall);
+            }
+            if programdata.get_lamports().saturating_add(buffer_lamports)
+                < programdata_balance_required
+            {
+                ic_logger_msg!(
+                    log_collector,
+                    "Buffer account balance too low to fund upgrade"
+                );
+                return Err(InstructionError::InsufficientFunds);
+            }
+            if let UpgradeableLoaderState::ProgramData {
+                slot: _,
+                upgrade_authority_address,
+            } = programdata.get_state()?
+            {
+                if upgrade_authority_address.is_none() {
+                    ic_logger_msg!(log_collector, "Program not upgradeable");
+                    return Err(InstructionError::Immutable);
+                }
+                if upgrade_authority_address != authority_key {
+                    ic_logger_msg!(log_collector, "Incorrect upgrade authority provided");
+                    return Err(InstructionError::IncorrectAuthority);
+                }
+                if !instruction_context.is_signer(first_instruction_account.saturating_add(6))? {
+                    ic_logger_msg!(log_collector, "Upgrade authority did not sign");
+                    return Err(InstructionError::MissingRequiredSignature);
+                }
+            } else {
+                ic_logger_msg!(log_collector, "Invalid ProgramData account");
+                return Err(InstructionError::InvalidAccountData);
+            }
+            drop(programdata);
+
+            // Load and verify the program bits
+            let executor = create_executor(
+                first_instruction_account.saturating_add(2),
+                buffer_data_offset,
+                invoke_context,
+                use_jit,
+                true,
+                invoke_context
+                    .feature_set
+                    .is_active(&disable_deploy_of_alloc_free_syscall::id()),
+            )?;
+            invoke_context.update_executor(&new_program_id, executor);
+
+            let transaction_context = &invoke_context.transaction_context;
+            let instruction_context = transaction_context.get_current_instruction_context()?;
+
+            // Update the ProgramData account, record the upgraded data, and zero
+            // the rest
+            let mut programdata =
+                instruction_context.try_borrow_instruction_account(transaction_context, 0)?;
+            {
+                programdata.set_state(&UpgradeableLoaderState::ProgramData {
+                    slot: clock.slot,
+                    upgrade_authority_address: authority_key,
+                })?;
+                let dst_slice = programdata
+                    .get_data_mut()?
+                    .get_mut(
+                        programdata_data_offset
+                            ..programdata_data_offset.saturating_add(buffer_data_len),
+                    )
+                    .ok_or(InstructionError::AccountDataTooSmall)?;
+                let buffer =
+                    instruction_context.try_borrow_instruction_account(transaction_context, 2)?;
+                let src_slice = buffer
+                    .get_data()
+                    .get(buffer_data_offset..)
+                    .ok_or(InstructionError::AccountDataTooSmall)?;
+                dst_slice.copy_from_slice(src_slice);
+            }
+            programdata
+                .get_data_mut()?
+                .get_mut(programdata_data_offset.saturating_add(buffer_data_len)..)
+                .ok_or(InstructionError::AccountDataTooSmall)?
+                .fill(0);
+
+            // Fund ProgramData to rent-exemption, spill the rest
+
+            let programdata_lamports = programdata.get_lamports();
+            programdata.set_lamports(programdata_balance_required)?;
+            drop(programdata);
+
+            let mut buffer =
+                instruction_context.try_borrow_instruction_account(transaction_context, 2)?;
+            buffer.set_lamports(0)?;
+            drop(buffer);
+
+            let mut spill =
+                instruction_context.try_borrow_instruction_account(transaction_context, 3)?;
+            spill.checked_add_lamports(
+                programdata_lamports
+                    .saturating_add(buffer_lamports)
+                    .saturating_sub(programdata_balance_required),
+            )?;
+
+            ic_logger_msg!(log_collector, "Upgraded program {:?}", new_program_id);
+        }
+        UpgradeableLoaderInstruction::SetAuthority => {
+            instruction_context.check_number_of_instruction_accounts(2)?;
+            let mut account =
+                instruction_context.try_borrow_instruction_account(transaction_context, 0)?;
+            let present_authority_key = transaction_context.get_key_of_account_at_index(
+                instruction_context
+                    .get_index_in_transaction(first_instruction_account.saturating_add(1))?,
+            )?;
+            let new_authority = instruction_context
+                .get_index_in_transaction(first_instruction_account.saturating_add(2))
+                .and_then(|index_in_transaction| {
+                    transaction_context.get_key_of_account_at_index(index_in_transaction)
+                })
+                .ok();
+
+            match account.get_state()? {
+                UpgradeableLoaderState::Buffer { authority_address } => {
+                    if new_authority.is_none() {
+                        ic_logger_msg!(log_collector, "Buffer authority is not optional");
+                        return Err(InstructionError::IncorrectAuthority);
+                    }
+                    if authority_address.is_none() {
+                        ic_logger_msg!(log_collector, "Buffer is immutable");
+                        return Err(InstructionError::Immutable);
+                    }
+                    if authority_address != Some(*present_authority_key) {
+                        ic_logger_msg!(log_collector, "Incorrect buffer authority provided");
+                        return Err(InstructionError::IncorrectAuthority);
+                    }
+                    if !instruction_context
+                        .is_signer(first_instruction_account.saturating_add(1))?
+                    {
+                        ic_logger_msg!(log_collector, "Buffer authority did not sign");
+                        return Err(InstructionError::MissingRequiredSignature);
+                    }
+                    account.set_state(&UpgradeableLoaderState::Buffer {
+                        authority_address: new_authority.cloned(),
+                    })?;
+                }
+                UpgradeableLoaderState::ProgramData {
+                    slot,
+                    upgrade_authority_address,
+                } => {
+                    if upgrade_authority_address.is_none() {
+                        ic_logger_msg!(log_collector, "Program not upgradeable");
+                        return Err(InstructionError::Immutable);
+                    }
+                    if upgrade_authority_address != Some(*present_authority_key) {
+                        ic_logger_msg!(log_collector, "Incorrect upgrade authority provided");
+                        return Err(InstructionError::IncorrectAuthority);
+                    }
+                    if !instruction_context
+                        .is_signer(first_instruction_account.saturating_add(1))?
+                    {
+                        ic_logger_msg!(log_collector, "Upgrade authority did not sign");
+                        return Err(InstructionError::MissingRequiredSignature);
+                    }
+                    account.set_state(&UpgradeableLoaderState::ProgramData {
+                        slot,
+                        upgrade_authority_address: new_authority.cloned(),
+                    })?;
+                }
+                _ => {
+                    ic_logger_msg!(log_collector, "Account does not support authorities");
+                    return Err(InstructionError::InvalidArgument);
+                }
+            }
+
+            ic_logger_msg!(log_collector, "New authority {:?}", new_authority);
+        }
+        UpgradeableLoaderInstruction::Close => {
+            instruction_context.check_number_of_instruction_accounts(2)?;
+            if instruction_context.get_index_in_transaction(first_instruction_account)?
+                == instruction_context
+                    .get_index_in_transaction(first_instruction_account.saturating_add(1))?
+            {
+                ic_logger_msg!(
+                    log_collector,
+                    "Recipient is the same as the account being closed"
+                );
+                return Err(InstructionError::InvalidArgument);
+            }
+            let mut close_account =
+                instruction_context.try_borrow_instruction_account(transaction_context, 0)?;
+            let close_key = *close_account.get_key();
+            match close_account.get_state()? {
+                UpgradeableLoaderState::Uninitialized => {
+                    let close_lamports = close_account.get_lamports();
+                    close_account.set_lamports(0)?;
+                    drop(close_account);
+                    let mut recipient_account = instruction_context
+                        .try_borrow_instruction_account(transaction_context, 1)?;
+                    recipient_account.checked_add_lamports(close_lamports)?;
+
+                    ic_logger_msg!(log_collector, "Closed Uninitialized {}", close_key);
+                }
+                UpgradeableLoaderState::Buffer { authority_address } => {
+                    instruction_context.check_number_of_instruction_accounts(3)?;
+                    drop(close_account);
+                    common_close_account(
+                        &authority_address,
+                        transaction_context,
+                        instruction_context,
+                        &log_collector,
+                    )?;
+
+                    ic_logger_msg!(log_collector, "Closed Buffer {}", close_key);
+                }
+                UpgradeableLoaderState::ProgramData {
+                    slot: _,
+                    upgrade_authority_address: authority_address,
+                } => {
+                    instruction_context.check_number_of_instruction_accounts(4)?;
+                    drop(close_account);
+                    let program_account = instruction_context
+                        .try_borrow_instruction_account(transaction_context, 3)?;
+                    let program_key = *program_account.get_key();
+
+                    if !program_account.is_writable() {
+                        ic_logger_msg!(log_collector, "Program account is not writable");
+                        return Err(InstructionError::InvalidArgument);
+                    }
+                    if program_account.get_owner() != program_id {
+                        ic_logger_msg!(log_collector, "Program account not owned by loader");
+                        return Err(InstructionError::IncorrectProgramId);
+                    }
+
+                    match program_account.get_state()? {
+                        UpgradeableLoaderState::Program {
+                            programdata_address,
+                        } => {
+                            if programdata_address != close_key {
+                                ic_logger_msg!(
+                                    log_collector,
+                                    "ProgramData account does not match ProgramData account"
+                                );
+                                return Err(InstructionError::InvalidArgument);
+                            }
+
+                            drop(program_account);
+                            common_close_account(
+                                &authority_address,
+                                transaction_context,
+                                instruction_context,
+                                &log_collector,
+                            )?;
+                        }
+                        _ => {
+                            ic_logger_msg!(log_collector, "Invalid Program account");
+                            return Err(InstructionError::InvalidArgument);
+                        }
+                    }
+
+                    ic_logger_msg!(log_collector, "Closed Program {}", program_key);
+                }
+                _ => {
+                    ic_logger_msg!(log_collector, "Account does not support closing");
+                    return Err(InstructionError::InvalidArgument);
+                }
+            }
+        }
+    }
+
+    Ok(())
+}
+
+fn common_close_account(
+    authority_address: &Option<Pubkey>,
+    transaction_context: &TransactionContext,
+    instruction_context: &InstructionContext,
+    log_collector: &Option<Rc<RefCell<LogCollector>>>,
+) -> Result<(), InstructionError> {
+    if authority_address.is_none() {
+        ic_logger_msg!(log_collector, "Account is immutable");
+        return Err(InstructionError::Immutable);
+    }
+    if *authority_address
+        != Some(*instruction_context.get_instruction_account_key(transaction_context, 2)?)
+    {
+        ic_logger_msg!(log_collector, "Incorrect authority provided");
+        return Err(InstructionError::IncorrectAuthority);
+    }
+    if !instruction_context.is_signer(
+        instruction_context
+            .get_number_of_program_accounts()
+            .saturating_add(2),
+    )? {
+        ic_logger_msg!(log_collector, "Authority did not sign");
+        return Err(InstructionError::MissingRequiredSignature);
+    }
+
+    let mut close_account =
+        instruction_context.try_borrow_instruction_account(transaction_context, 0)?;
+    let mut recipient_account =
+        instruction_context.try_borrow_instruction_account(transaction_context, 1)?;
+    recipient_account.checked_add_lamports(close_account.get_lamports())?;
+    close_account.set_lamports(0)?;
+    close_account.set_state(&UpgradeableLoaderState::Uninitialized)?;
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use {
+        super::*,
+        solana_program_runtime::invoke_context::mock_process_instruction,
+        solana_runtime::{bank::Bank, bank_client::BankClient},
+        solana_sdk::{
+            account::{
+                create_account_shared_data_for_test, AccountSharedData, ReadableAccount,
+                WritableAccount,
+            },
+            account_utils::StateMut,
+            bpf_loader_upgradeable,
+            client::SyncClient,
+            clock::Clock,
+            feature_set::FeatureSet,
+            genesis_config::create_genesis_config,
+            instruction::Instruction,
+            message::Message,
+            native_token::LAMPORTS_PER_SOL,
+            rent::Rent,
+            signature::{Keypair, Signer},
+            system_program, sysvar,
+            transaction::TransactionError,
+        },
+        std::{fs::File, io::Read, sync::Arc},
+    };
+
+    fn process_instruction_without_jit(
+        first_instruction_account: usize,
+        invoke_context: &mut InvokeContext,
+    ) -> Result<(), InstructionError> {
+        super::process_instruction(first_instruction_account, invoke_context, false)
+    }
+
+    fn test_process_instruction(
+        loader_id: &Pubkey,
+        program_indices: &[usize],
+        instruction_data: &[u8],
+        transaction_accounts: Vec<(Pubkey, AccountSharedData)>,
+        instruction_accounts: Vec<AccountMeta>,
+        expected_result: Result<(), InstructionError>,
+    ) -> Vec<AccountSharedData> {
+        mock_process_instruction(
+            loader_id,
+            program_indices.to_vec(),
+            instruction_data,
+            transaction_accounts,
+            instruction_accounts,
+            None,
+            None,
+            expected_result,
+            process_instruction_without_jit,
+        )
+    }
+
+    #[test]
+    fn test_initialize_buffer() {
+        let loader_id = bpf_loader_upgradeable::id();
+        let buffer_address = Pubkey::new_unique();
+        let buffer_account =
+            AccountSharedData::new(1, UpgradeableLoaderState::size_of_buffer(9), &loader_id);
+        let authority_address = Pubkey::new_unique();
+        let authority_account =
+            AccountSharedData::new(1, UpgradeableLoaderState::size_of_buffer(9), &loader_id);
+        let instruction_data =
+            bincode::serialize(&UpgradeableLoaderInstruction::InitializeBuffer).unwrap();
+        let instruction_accounts = vec![
+            AccountMeta {
+                pubkey: buffer_address,
+                is_signer: false,
+                is_writable: false,
+            },
+            AccountMeta {
+                pubkey: authority_address,
+                is_signer: false,
+                is_writable: false,
+            },
+        ];
+
+        // Case: Success
+        let accounts = test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction_data,
+            vec![
+                (buffer_address, buffer_account),
+                (authority_address, authority_account),
+            ],
+            instruction_accounts.clone(),
+            Ok(()),
+        );
+        let state: UpgradeableLoaderState = accounts.first().unwrap().state().unwrap();
+        assert_eq!(
+            state,
+            UpgradeableLoaderState::Buffer {
+                authority_address: Some(authority_address)
+            }
+        );
+
+        // Case: Already initialized
+        let accounts = test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction_data,
+            vec![
+                (buffer_address, accounts.first().unwrap().clone()),
+                (authority_address, accounts.get(1).unwrap().clone()),
+            ],
+            instruction_accounts,
+            Err(InstructionError::AccountAlreadyInitialized),
+        );
+        let state: UpgradeableLoaderState = accounts.first().unwrap().state().unwrap();
+        assert_eq!(
+            state,
+            UpgradeableLoaderState::Buffer {
+                authority_address: Some(authority_address)
+            }
+        );
+    }
+
+    #[test]
+    fn test_write() {
+        let loader_id = bpf_loader_upgradeable::id();
+        let buffer_address = Pubkey::new_unique();
+        let mut buffer_account =
+            AccountSharedData::new(1, UpgradeableLoaderState::size_of_buffer(9), &loader_id);
+        let instruction_accounts = vec![
+            AccountMeta {
+                pubkey: buffer_address,
+                is_signer: false,
+                is_writable: false,
+            },
+            AccountMeta {
+                pubkey: buffer_address,
+                is_signer: true,
+                is_writable: false,
+            },
+        ];
+
+        // Case: Not initialized
+        let instruction = bincode::serialize(&UpgradeableLoaderInstruction::Write {
+            offset: 0,
+            bytes: vec![42; 9],
+        })
+        .unwrap();
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![(buffer_address, buffer_account.clone())],
+            instruction_accounts.clone(),
+            Err(InstructionError::InvalidAccountData),
+        );
+
+        // Case: Write entire buffer
+        let instruction = bincode::serialize(&UpgradeableLoaderInstruction::Write {
+            offset: 0,
+            bytes: vec![42; 9],
+        })
+        .unwrap();
+        buffer_account
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: Some(buffer_address),
+            })
+            .unwrap();
+        let accounts = test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![(buffer_address, buffer_account.clone())],
+            instruction_accounts.clone(),
+            Ok(()),
+        );
+        let state: UpgradeableLoaderState = accounts.first().unwrap().state().unwrap();
+        assert_eq!(
+            state,
+            UpgradeableLoaderState::Buffer {
+                authority_address: Some(buffer_address)
+            }
+        );
+        assert_eq!(
+            &accounts
+                .first()
+                .unwrap()
+                .data()
+                .get(UpgradeableLoaderState::size_of_buffer_metadata()..)
+                .unwrap(),
+            &[42; 9]
+        );
+
+        // Case: Write portion of the buffer
+        let instruction = bincode::serialize(&UpgradeableLoaderInstruction::Write {
+            offset: 3,
+            bytes: vec![42; 6],
+        })
+        .unwrap();
+        let mut buffer_account =
+            AccountSharedData::new(1, UpgradeableLoaderState::size_of_buffer(9), &loader_id);
+        buffer_account
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: Some(buffer_address),
+            })
+            .unwrap();
+        let accounts = test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![(buffer_address, buffer_account.clone())],
+            instruction_accounts.clone(),
+            Ok(()),
+        );
+        let state: UpgradeableLoaderState = accounts.first().unwrap().state().unwrap();
+        assert_eq!(
+            state,
+            UpgradeableLoaderState::Buffer {
+                authority_address: Some(buffer_address)
+            }
+        );
+        assert_eq!(
+            &accounts
+                .first()
+                .unwrap()
+                .data()
+                .get(UpgradeableLoaderState::size_of_buffer_metadata()..)
+                .unwrap(),
+            &[0, 0, 0, 42, 42, 42, 42, 42, 42]
+        );
+
+        // Case: overflow size
+        let instruction = bincode::serialize(&UpgradeableLoaderInstruction::Write {
+            offset: 0,
+            bytes: vec![42; 10],
+        })
+        .unwrap();
+        buffer_account
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: Some(buffer_address),
+            })
+            .unwrap();
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![(buffer_address, buffer_account.clone())],
+            instruction_accounts.clone(),
+            Err(InstructionError::AccountDataTooSmall),
+        );
+
+        // Case: overflow offset
+        let instruction = bincode::serialize(&UpgradeableLoaderInstruction::Write {
+            offset: 1,
+            bytes: vec![42; 9],
+        })
+        .unwrap();
+        buffer_account
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: Some(buffer_address),
+            })
+            .unwrap();
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![(buffer_address, buffer_account.clone())],
+            instruction_accounts.clone(),
+            Err(InstructionError::AccountDataTooSmall),
+        );
+
+        // Case: Not signed
+        let instruction = bincode::serialize(&UpgradeableLoaderInstruction::Write {
+            offset: 0,
+            bytes: vec![42; 9],
+        })
+        .unwrap();
+        buffer_account
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: Some(buffer_address),
+            })
+            .unwrap();
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![(buffer_address, buffer_account.clone())],
+            vec![
+                AccountMeta {
+                    pubkey: buffer_address,
+                    is_signer: false,
+                    is_writable: false,
+                },
+                AccountMeta {
+                    pubkey: buffer_address,
+                    is_signer: false,
+                    is_writable: false,
+                },
+            ],
+            Err(InstructionError::MissingRequiredSignature),
+        );
+
+        // Case: wrong authority
+        let authority_address = Pubkey::new_unique();
+        let instruction = bincode::serialize(&UpgradeableLoaderInstruction::Write {
+            offset: 1,
+            bytes: vec![42; 9],
+        })
+        .unwrap();
+        buffer_account
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: Some(buffer_address),
+            })
+            .unwrap();
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![
+                (buffer_address, buffer_account.clone()),
+                (authority_address, buffer_account.clone()),
+            ],
+            vec![
+                AccountMeta {
+                    pubkey: buffer_address,
+                    is_signer: false,
+                    is_writable: false,
+                },
+                AccountMeta {
+                    pubkey: authority_address,
+                    is_signer: false,
+                    is_writable: false,
+                },
+            ],
+            Err(InstructionError::IncorrectAuthority),
+        );
+
+        // Case: None authority
+        let instruction = bincode::serialize(&UpgradeableLoaderInstruction::Write {
+            offset: 1,
+            bytes: vec![42; 9],
+        })
+        .unwrap();
+        buffer_account
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: None,
+            })
+            .unwrap();
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![(buffer_address, buffer_account.clone())],
+            instruction_accounts,
+            Err(InstructionError::Immutable),
+        );
+    }
+
+    fn truncate_data(account: &mut AccountSharedData, len: usize) {
+        let mut data = account.data().to_vec();
+        data.truncate(len);
+        account.set_data(data);
+    }
+
+    #[test]
+    fn test_deploy_with_max_len() {
+        let (genesis_config, mint_keypair) = create_genesis_config(1_000_000_000);
+        let mut bank = Bank::new_for_tests(&genesis_config);
+        bank.feature_set = Arc::new(FeatureSet::all_enabled());
+        bank.add_builtin(
+            "solana_bpf_loader_upgradeable_program",
+            &bpf_loader_upgradeable::id(),
+            process_instruction_without_jit,
+        );
+        let bank = Arc::new(bank);
+        let bank_client = BankClient::new_shared(&bank);
+
+        // Setup keypairs and addresses
+        let payer_keypair = Keypair::new();
+        let program_keypair = Keypair::new();
+        let buffer_address = Pubkey::new_unique();
+        let (programdata_address, _) = Pubkey::find_program_address(
+            &[program_keypair.pubkey().as_ref()],
+            &bpf_loader_upgradeable::id(),
+        );
+        let upgrade_authority_keypair = Keypair::new();
+
+        // Load program file
+        let mut file = File::open("test_elfs/out/noop_aligned.so").expect("file open failed");
+        let mut elf = Vec::new();
+        file.read_to_end(&mut elf).unwrap();
+
+        // Compute rent exempt balances
+        let program_len = elf.len();
+        let min_program_balance =
+            bank.get_minimum_balance_for_rent_exemption(UpgradeableLoaderState::size_of_program());
+        let min_buffer_balance = bank.get_minimum_balance_for_rent_exemption(
+            UpgradeableLoaderState::size_of_buffer(program_len),
+        );
+        let min_programdata_balance = bank.get_minimum_balance_for_rent_exemption(
+            UpgradeableLoaderState::size_of_programdata(program_len),
+        );
+
+        // Setup accounts
+        let buffer_account = {
+            let mut account = AccountSharedData::new(
+                min_buffer_balance,
+                UpgradeableLoaderState::size_of_buffer(elf.len()),
+                &bpf_loader_upgradeable::id(),
+            );
+            account
+                .set_state(&UpgradeableLoaderState::Buffer {
+                    authority_address: Some(upgrade_authority_keypair.pubkey()),
+                })
+                .unwrap();
+            account
+                .data_as_mut_slice()
+                .get_mut(UpgradeableLoaderState::size_of_buffer_metadata()..)
+                .unwrap()
+                .copy_from_slice(&elf);
+            account
+        };
+        let program_account = AccountSharedData::new(
+            min_programdata_balance,
+            UpgradeableLoaderState::size_of_program(),
+            &bpf_loader_upgradeable::id(),
+        );
+        let programdata_account = AccountSharedData::new(
+            1,
+            UpgradeableLoaderState::size_of_programdata(elf.len()),
+            &bpf_loader_upgradeable::id(),
+        );
+
+        // Test successful deploy
+        let payer_base_balance = LAMPORTS_PER_SOL;
+        let deploy_fees = {
+            let fee_calculator = genesis_config.fee_rate_governor.create_fee_calculator();
+            3 * fee_calculator.lamports_per_signature
+        };
+        let min_payer_balance = min_program_balance
+            .saturating_add(min_programdata_balance)
+            .saturating_sub(min_buffer_balance.saturating_add(deploy_fees));
+        bank.store_account(
+            &payer_keypair.pubkey(),
+            &AccountSharedData::new(
+                payer_base_balance.saturating_add(min_payer_balance),
+                0,
+                &system_program::id(),
+            ),
+        );
+        bank.store_account(&buffer_address, &buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        bank.store_account(&programdata_address, &AccountSharedData::default());
+        let message = Message::new(
+            &bpf_loader_upgradeable::deploy_with_max_program_len(
+                &payer_keypair.pubkey(),
+                &program_keypair.pubkey(),
+                &buffer_address,
+                &upgrade_authority_keypair.pubkey(),
+                min_program_balance,
+                elf.len(),
+            )
+            .unwrap(),
+            Some(&payer_keypair.pubkey()),
+        );
+        assert!(bank_client
+            .send_and_confirm_message(
+                &[&payer_keypair, &program_keypair, &upgrade_authority_keypair],
+                message
+            )
+            .is_ok());
+        assert_eq!(
+            bank.get_balance(&payer_keypair.pubkey()),
+            payer_base_balance
+        );
+        assert_eq!(bank.get_balance(&buffer_address), 0);
+        assert_eq!(None, bank.get_account(&buffer_address));
+        let post_program_account = bank.get_account(&program_keypair.pubkey()).unwrap();
+        assert_eq!(post_program_account.lamports(), min_program_balance);
+        assert_eq!(post_program_account.owner(), &bpf_loader_upgradeable::id());
+        assert_eq!(
+            post_program_account.data().len(),
+            UpgradeableLoaderState::size_of_program()
+        );
+        let state: UpgradeableLoaderState = post_program_account.state().unwrap();
+        assert_eq!(
+            state,
+            UpgradeableLoaderState::Program {
+                programdata_address
+            }
+        );
+        let post_programdata_account = bank.get_account(&programdata_address).unwrap();
+        assert_eq!(post_programdata_account.lamports(), min_programdata_balance);
+        assert_eq!(
+            post_programdata_account.owner(),
+            &bpf_loader_upgradeable::id()
+        );
+        let state: UpgradeableLoaderState = post_programdata_account.state().unwrap();
+        assert_eq!(
+            state,
+            UpgradeableLoaderState::ProgramData {
+                slot: bank_client.get_slot().unwrap(),
+                upgrade_authority_address: Some(upgrade_authority_keypair.pubkey())
+            }
+        );
+        for (i, byte) in post_programdata_account
+            .data()
+            .get(UpgradeableLoaderState::size_of_programdata_metadata()..)
+            .unwrap()
+            .iter()
+            .enumerate()
+        {
+            assert_eq!(*elf.get(i).unwrap(), *byte);
+        }
+
+        // Invoke deployed program
+        test_process_instruction(
+            &bpf_loader_upgradeable::id(),
+            &[0, 1],
+            &[],
+            vec![
+                (programdata_address, post_programdata_account),
+                (program_keypair.pubkey(), post_program_account),
+            ],
+            Vec::new(),
+            Ok(()),
+        );
+
+        // Test initialized program account
+        bank.clear_signatures();
+        bank.store_account(&buffer_address, &buffer_account);
+        let message = Message::new(
+            &[Instruction::new_with_bincode(
+                bpf_loader_upgradeable::id(),
+                &UpgradeableLoaderInstruction::DeployWithMaxDataLen {
+                    max_data_len: elf.len(),
+                },
+                vec![
+                    AccountMeta::new(mint_keypair.pubkey(), true),
+                    AccountMeta::new(programdata_address, false),
+                    AccountMeta::new(program_keypair.pubkey(), false),
+                    AccountMeta::new(buffer_address, false),
+                    AccountMeta::new_readonly(sysvar::rent::id(), false),
+                    AccountMeta::new_readonly(sysvar::clock::id(), false),
+                    AccountMeta::new_readonly(system_program::id(), false),
+                    AccountMeta::new_readonly(upgrade_authority_keypair.pubkey(), true),
+                ],
+            )],
+            Some(&mint_keypair.pubkey()),
+        );
+        assert_eq!(
+            TransactionError::InstructionError(0, InstructionError::AccountAlreadyInitialized),
+            bank_client
+                .send_and_confirm_message(&[&mint_keypair, &upgrade_authority_keypair], message)
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Test initialized ProgramData account
+        bank.clear_signatures();
+        bank.store_account(&buffer_address, &buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        let message = Message::new(
+            &bpf_loader_upgradeable::deploy_with_max_program_len(
+                &mint_keypair.pubkey(),
+                &program_keypair.pubkey(),
+                &buffer_address,
+                &upgrade_authority_keypair.pubkey(),
+                min_program_balance,
+                elf.len(),
+            )
+            .unwrap(),
+            Some(&mint_keypair.pubkey()),
+        );
+        assert_eq!(
+            TransactionError::InstructionError(1, InstructionError::Custom(0)),
+            bank_client
+                .send_and_confirm_message(
+                    &[&mint_keypair, &program_keypair, &upgrade_authority_keypair],
+                    message
+                )
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Test deploy no authority
+        bank.clear_signatures();
+        bank.store_account(&buffer_address, &buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &program_account);
+        bank.store_account(&programdata_address, &programdata_account);
+        let message = Message::new(
+            &[Instruction::new_with_bincode(
+                bpf_loader_upgradeable::id(),
+                &UpgradeableLoaderInstruction::DeployWithMaxDataLen {
+                    max_data_len: elf.len(),
+                },
+                vec![
+                    AccountMeta::new(mint_keypair.pubkey(), true),
+                    AccountMeta::new(programdata_address, false),
+                    AccountMeta::new(program_keypair.pubkey(), false),
+                    AccountMeta::new(buffer_address, false),
+                    AccountMeta::new_readonly(sysvar::rent::id(), false),
+                    AccountMeta::new_readonly(sysvar::clock::id(), false),
+                    AccountMeta::new_readonly(system_program::id(), false),
+                ],
+            )],
+            Some(&mint_keypair.pubkey()),
+        );
+        assert_eq!(
+            TransactionError::InstructionError(0, InstructionError::NotEnoughAccountKeys),
+            bank_client
+                .send_and_confirm_message(&[&mint_keypair], message)
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Test deploy authority not a signer
+        bank.clear_signatures();
+        bank.store_account(&buffer_address, &buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &program_account);
+        bank.store_account(&programdata_address, &programdata_account);
+        let message = Message::new(
+            &[Instruction::new_with_bincode(
+                bpf_loader_upgradeable::id(),
+                &UpgradeableLoaderInstruction::DeployWithMaxDataLen {
+                    max_data_len: elf.len(),
+                },
+                vec![
+                    AccountMeta::new(mint_keypair.pubkey(), true),
+                    AccountMeta::new(programdata_address, false),
+                    AccountMeta::new(program_keypair.pubkey(), false),
+                    AccountMeta::new(buffer_address, false),
+                    AccountMeta::new_readonly(sysvar::rent::id(), false),
+                    AccountMeta::new_readonly(sysvar::clock::id(), false),
+                    AccountMeta::new_readonly(system_program::id(), false),
+                    AccountMeta::new_readonly(upgrade_authority_keypair.pubkey(), false),
+                ],
+            )],
+            Some(&mint_keypair.pubkey()),
+        );
+        assert_eq!(
+            TransactionError::InstructionError(0, InstructionError::MissingRequiredSignature),
+            bank_client
+                .send_and_confirm_message(&[&mint_keypair], message)
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Test invalid Buffer account state
+        bank.clear_signatures();
+        bank.store_account(&buffer_address, &AccountSharedData::default());
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        bank.store_account(&programdata_address, &AccountSharedData::default());
+        let message = Message::new(
+            &bpf_loader_upgradeable::deploy_with_max_program_len(
+                &mint_keypair.pubkey(),
+                &program_keypair.pubkey(),
+                &buffer_address,
+                &upgrade_authority_keypair.pubkey(),
+                min_program_balance,
+                elf.len(),
+            )
+            .unwrap(),
+            Some(&mint_keypair.pubkey()),
+        );
+        assert_eq!(
+            TransactionError::InstructionError(1, InstructionError::InvalidAccountData),
+            bank_client
+                .send_and_confirm_message(
+                    &[&mint_keypair, &program_keypair, &upgrade_authority_keypair],
+                    message
+                )
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Test program account not rent exempt
+        bank.clear_signatures();
+        bank.store_account(&buffer_address, &buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        bank.store_account(&programdata_address, &AccountSharedData::default());
+        let message = Message::new(
+            &bpf_loader_upgradeable::deploy_with_max_program_len(
+                &mint_keypair.pubkey(),
+                &program_keypair.pubkey(),
+                &buffer_address,
+                &upgrade_authority_keypair.pubkey(),
+                min_program_balance.saturating_sub(1),
+                elf.len(),
+            )
+            .unwrap(),
+            Some(&mint_keypair.pubkey()),
+        );
+        assert_eq!(
+            TransactionError::InstructionError(1, InstructionError::ExecutableAccountNotRentExempt),
+            bank_client
+                .send_and_confirm_message(
+                    &[&mint_keypair, &program_keypair, &upgrade_authority_keypair],
+                    message
+                )
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Test program account not rent exempt because data is larger than needed
+        bank.clear_signatures();
+        bank.store_account(&buffer_address, &buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        bank.store_account(&programdata_address, &AccountSharedData::default());
+        let mut instructions = bpf_loader_upgradeable::deploy_with_max_program_len(
+            &mint_keypair.pubkey(),
+            &program_keypair.pubkey(),
+            &buffer_address,
+            &upgrade_authority_keypair.pubkey(),
+            min_program_balance,
+            elf.len(),
+        )
+        .unwrap();
+        *instructions.get_mut(0).unwrap() = system_instruction::create_account(
+            &mint_keypair.pubkey(),
+            &program_keypair.pubkey(),
+            min_program_balance,
+            (UpgradeableLoaderState::size_of_program() as u64).saturating_add(1),
+            &bpf_loader_upgradeable::id(),
+        );
+        let message = Message::new(&instructions, Some(&mint_keypair.pubkey()));
+        assert_eq!(
+            TransactionError::InstructionError(1, InstructionError::ExecutableAccountNotRentExempt),
+            bank_client
+                .send_and_confirm_message(
+                    &[&mint_keypair, &program_keypair, &upgrade_authority_keypair],
+                    message
+                )
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Test program account too small
+        bank.clear_signatures();
+        bank.store_account(&buffer_address, &buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        bank.store_account(&programdata_address, &AccountSharedData::default());
+        let mut instructions = bpf_loader_upgradeable::deploy_with_max_program_len(
+            &mint_keypair.pubkey(),
+            &program_keypair.pubkey(),
+            &buffer_address,
+            &upgrade_authority_keypair.pubkey(),
+            min_program_balance,
+            elf.len(),
+        )
+        .unwrap();
+        *instructions.get_mut(0).unwrap() = system_instruction::create_account(
+            &mint_keypair.pubkey(),
+            &program_keypair.pubkey(),
+            min_program_balance,
+            (UpgradeableLoaderState::size_of_program() as u64).saturating_sub(1),
+            &bpf_loader_upgradeable::id(),
+        );
+        let message = Message::new(&instructions, Some(&mint_keypair.pubkey()));
+        assert_eq!(
+            TransactionError::InstructionError(1, InstructionError::AccountDataTooSmall),
+            bank_client
+                .send_and_confirm_message(
+                    &[&mint_keypair, &program_keypair, &upgrade_authority_keypair],
+                    message
+                )
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Test Insufficient payer funds (need more funds to cover the
+        // difference between buffer lamports and programdata lamports)
+        bank.clear_signatures();
+        bank.store_account(
+            &mint_keypair.pubkey(),
+            &AccountSharedData::new(
+                deploy_fees.saturating_add(min_program_balance),
+                0,
+                &system_program::id(),
+            ),
+        );
+        bank.store_account(&buffer_address, &buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        bank.store_account(&programdata_address, &AccountSharedData::default());
+        let message = Message::new(
+            &bpf_loader_upgradeable::deploy_with_max_program_len(
+                &mint_keypair.pubkey(),
+                &program_keypair.pubkey(),
+                &buffer_address,
+                &upgrade_authority_keypair.pubkey(),
+                min_program_balance,
+                elf.len(),
+            )
+            .unwrap(),
+            Some(&mint_keypair.pubkey()),
+        );
+        assert_eq!(
+            TransactionError::InstructionError(1, InstructionError::Custom(1)),
+            bank_client
+                .send_and_confirm_message(
+                    &[&mint_keypair, &program_keypair, &upgrade_authority_keypair],
+                    message
+                )
+                .unwrap_err()
+                .unwrap()
+        );
+        bank.store_account(
+            &mint_keypair.pubkey(),
+            &AccountSharedData::new(1_000_000_000, 0, &system_program::id()),
+        );
+
+        // Test max_data_len
+        bank.clear_signatures();
+        bank.store_account(&buffer_address, &buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        bank.store_account(&programdata_address, &AccountSharedData::default());
+        let message = Message::new(
+            &bpf_loader_upgradeable::deploy_with_max_program_len(
+                &mint_keypair.pubkey(),
+                &program_keypair.pubkey(),
+                &buffer_address,
+                &upgrade_authority_keypair.pubkey(),
+                min_program_balance,
+                elf.len().saturating_sub(1),
+            )
+            .unwrap(),
+            Some(&mint_keypair.pubkey()),
+        );
+        assert_eq!(
+            TransactionError::InstructionError(1, InstructionError::AccountDataTooSmall),
+            bank_client
+                .send_and_confirm_message(
+                    &[&mint_keypair, &program_keypair, &upgrade_authority_keypair],
+                    message
+                )
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Test max_data_len too large
+        bank.clear_signatures();
+        bank.store_account(
+            &mint_keypair.pubkey(),
+            &AccountSharedData::new(u64::MAX / 2, 0, &system_program::id()),
+        );
+        let mut modified_buffer_account = buffer_account.clone();
+        modified_buffer_account.set_lamports(u64::MAX / 2);
+        bank.store_account(&buffer_address, &modified_buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        bank.store_account(&programdata_address, &AccountSharedData::default());
+        let message = Message::new(
+            &bpf_loader_upgradeable::deploy_with_max_program_len(
+                &mint_keypair.pubkey(),
+                &program_keypair.pubkey(),
+                &buffer_address,
+                &upgrade_authority_keypair.pubkey(),
+                min_program_balance,
+                usize::MAX,
+            )
+            .unwrap(),
+            Some(&mint_keypair.pubkey()),
+        );
+        assert_eq!(
+            TransactionError::InstructionError(1, InstructionError::InvalidArgument),
+            bank_client
+                .send_and_confirm_message(
+                    &[&mint_keypair, &program_keypair, &upgrade_authority_keypair],
+                    message
+                )
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Test not the system account
+        bank.clear_signatures();
+        bank.store_account(&buffer_address, &buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        bank.store_account(&programdata_address, &AccountSharedData::default());
+        let mut instructions = bpf_loader_upgradeable::deploy_with_max_program_len(
+            &mint_keypair.pubkey(),
+            &program_keypair.pubkey(),
+            &buffer_address,
+            &upgrade_authority_keypair.pubkey(),
+            min_program_balance,
+            elf.len(),
+        )
+        .unwrap();
+        *instructions
+            .get_mut(1)
+            .unwrap()
+            .accounts
+            .get_mut(6)
+            .unwrap() = AccountMeta::new_readonly(Pubkey::new_unique(), false);
+        let message = Message::new(&instructions, Some(&mint_keypair.pubkey()));
+        assert_eq!(
+            TransactionError::InstructionError(1, InstructionError::MissingAccount),
+            bank_client
+                .send_and_confirm_message(
+                    &[&mint_keypair, &program_keypair, &upgrade_authority_keypair],
+                    message
+                )
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Test Bad ELF data
+        bank.clear_signatures();
+        let mut modified_buffer_account = buffer_account;
+        truncate_data(
+            &mut modified_buffer_account,
+            UpgradeableLoaderState::size_of_buffer(1),
+        );
+        bank.store_account(&buffer_address, &modified_buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        bank.store_account(&programdata_address, &AccountSharedData::default());
+        let message = Message::new(
+            &bpf_loader_upgradeable::deploy_with_max_program_len(
+                &mint_keypair.pubkey(),
+                &program_keypair.pubkey(),
+                &buffer_address,
+                &upgrade_authority_keypair.pubkey(),
+                min_program_balance,
+                elf.len(),
+            )
+            .unwrap(),
+            Some(&mint_keypair.pubkey()),
+        );
+        assert_eq!(
+            TransactionError::InstructionError(1, InstructionError::InvalidAccountData),
+            bank_client
+                .send_and_confirm_message(
+                    &[&mint_keypair, &program_keypair, &upgrade_authority_keypair],
+                    message
+                )
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Test small buffer account
+        bank.clear_signatures();
+        let mut modified_buffer_account = AccountSharedData::new(
+            min_programdata_balance,
+            UpgradeableLoaderState::size_of_buffer(elf.len()),
+            &bpf_loader_upgradeable::id(),
+        );
+        modified_buffer_account
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: Some(upgrade_authority_keypair.pubkey()),
+            })
+            .unwrap();
+        modified_buffer_account
+            .data_as_mut_slice()
+            .get_mut(UpgradeableLoaderState::size_of_buffer_metadata()..)
+            .unwrap()
+            .copy_from_slice(&elf);
+        truncate_data(&mut modified_buffer_account, 5);
+        bank.store_account(&buffer_address, &modified_buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        bank.store_account(&programdata_address, &AccountSharedData::default());
+        let message = Message::new(
+            &bpf_loader_upgradeable::deploy_with_max_program_len(
+                &mint_keypair.pubkey(),
+                &program_keypair.pubkey(),
+                &buffer_address,
+                &upgrade_authority_keypair.pubkey(),
+                min_program_balance,
+                elf.len(),
+            )
+            .unwrap(),
+            Some(&mint_keypair.pubkey()),
+        );
+        assert_eq!(
+            TransactionError::InstructionError(1, InstructionError::InvalidAccountData),
+            bank_client
+                .send_and_confirm_message(
+                    &[&mint_keypair, &program_keypair, &upgrade_authority_keypair],
+                    message
+                )
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Mismatched buffer and program authority
+        bank.clear_signatures();
+        let mut modified_buffer_account = AccountSharedData::new(
+            min_programdata_balance,
+            UpgradeableLoaderState::size_of_buffer(elf.len()),
+            &bpf_loader_upgradeable::id(),
+        );
+        modified_buffer_account
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: Some(buffer_address),
+            })
+            .unwrap();
+        modified_buffer_account
+            .data_as_mut_slice()
+            .get_mut(UpgradeableLoaderState::size_of_buffer_metadata()..)
+            .unwrap()
+            .copy_from_slice(&elf);
+        bank.store_account(&buffer_address, &modified_buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        bank.store_account(&programdata_address, &AccountSharedData::default());
+        let message = Message::new(
+            &bpf_loader_upgradeable::deploy_with_max_program_len(
+                &mint_keypair.pubkey(),
+                &program_keypair.pubkey(),
+                &buffer_address,
+                &upgrade_authority_keypair.pubkey(),
+                min_program_balance,
+                elf.len(),
+            )
+            .unwrap(),
+            Some(&mint_keypair.pubkey()),
+        );
+        assert_eq!(
+            TransactionError::InstructionError(1, InstructionError::IncorrectAuthority),
+            bank_client
+                .send_and_confirm_message(
+                    &[&mint_keypair, &program_keypair, &upgrade_authority_keypair],
+                    message
+                )
+                .unwrap_err()
+                .unwrap()
+        );
+
+        // Deploy buffer with mismatched None authority
+        bank.clear_signatures();
+        let mut modified_buffer_account = AccountSharedData::new(
+            min_programdata_balance,
+            UpgradeableLoaderState::size_of_buffer(elf.len()),
+            &bpf_loader_upgradeable::id(),
+        );
+        modified_buffer_account
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: None,
+            })
+            .unwrap();
+        modified_buffer_account
+            .data_as_mut_slice()
+            .get_mut(UpgradeableLoaderState::size_of_buffer_metadata()..)
+            .unwrap()
+            .copy_from_slice(&elf);
+        bank.store_account(&buffer_address, &modified_buffer_account);
+        bank.store_account(&program_keypair.pubkey(), &AccountSharedData::default());
+        bank.store_account(&programdata_address, &AccountSharedData::default());
+        let message = Message::new(
+            &bpf_loader_upgradeable::deploy_with_max_program_len(
+                &mint_keypair.pubkey(),
+                &program_keypair.pubkey(),
+                &buffer_address,
+                &upgrade_authority_keypair.pubkey(),
+                min_program_balance,
+                elf.len(),
+            )
+            .unwrap(),
+            Some(&mint_keypair.pubkey()),
+        );
+        assert_eq!(
+            TransactionError::InstructionError(1, InstructionError::IncorrectAuthority),
+            bank_client
+                .send_and_confirm_message(
+                    &[&mint_keypair, &program_keypair, &upgrade_authority_keypair],
+                    message
+                )
+                .unwrap_err()
+                .unwrap()
+        );
+    }
+
+    #[test]
+    fn test_upgrade() {
+        let mut file = File::open("test_elfs/out/noop_aligned.so").expect("file open failed");
+        let mut elf_orig = Vec::new();
+        file.read_to_end(&mut elf_orig).unwrap();
+        let mut file = File::open("test_elfs/out/noop_unaligned.so").expect("file open failed");
+        let mut elf_new = Vec::new();
+        file.read_to_end(&mut elf_new).unwrap();
+        assert_ne!(elf_orig.len(), elf_new.len());
+        const SLOT: u64 = 42;
+        let buffer_address = Pubkey::new_unique();
+        let upgrade_authority_address = Pubkey::new_unique();
+
+        fn get_accounts(
+            buffer_address: &Pubkey,
+            buffer_authority: &Pubkey,
+            upgrade_authority_address: &Pubkey,
+            elf_orig: &[u8],
+            elf_new: &[u8],
+        ) -> (Vec<(Pubkey, AccountSharedData)>, Vec<AccountMeta>) {
+            let loader_id = bpf_loader_upgradeable::id();
+            let program_address = Pubkey::new_unique();
+            let spill_address = Pubkey::new_unique();
+            let rent = Rent::default();
+            let min_program_balance =
+                1.max(rent.minimum_balance(UpgradeableLoaderState::size_of_program()));
+            let min_programdata_balance = 1.max(rent.minimum_balance(
+                UpgradeableLoaderState::size_of_programdata(elf_orig.len().max(elf_new.len())),
+            ));
+            let (programdata_address, _) =
+                Pubkey::find_program_address(&[program_address.as_ref()], &loader_id);
+            let mut buffer_account = AccountSharedData::new(
+                1,
+                UpgradeableLoaderState::size_of_buffer(elf_new.len()),
+                &bpf_loader_upgradeable::id(),
+            );
+            buffer_account
+                .set_state(&UpgradeableLoaderState::Buffer {
+                    authority_address: Some(*buffer_authority),
+                })
+                .unwrap();
+            buffer_account
+                .data_as_mut_slice()
+                .get_mut(UpgradeableLoaderState::size_of_buffer_metadata()..)
+                .unwrap()
+                .copy_from_slice(elf_new);
+            let mut programdata_account = AccountSharedData::new(
+                min_programdata_balance,
+                UpgradeableLoaderState::size_of_programdata(elf_orig.len().max(elf_new.len())),
+                &bpf_loader_upgradeable::id(),
+            );
+            programdata_account
+                .set_state(&UpgradeableLoaderState::ProgramData {
+                    slot: SLOT,
+                    upgrade_authority_address: Some(*upgrade_authority_address),
+                })
+                .unwrap();
+            let mut program_account = AccountSharedData::new(
+                min_program_balance,
+                UpgradeableLoaderState::size_of_program(),
+                &bpf_loader_upgradeable::id(),
+            );
+            program_account.set_executable(true);
+            program_account
+                .set_state(&UpgradeableLoaderState::Program {
+                    programdata_address,
+                })
+                .unwrap();
+            let spill_account = AccountSharedData::new(0, 0, &Pubkey::new_unique());
+            let rent_account = create_account_shared_data_for_test(&rent);
+            let clock_account = create_account_shared_data_for_test(&Clock {
+                slot: SLOT,
+                ..Clock::default()
+            });
+            let upgrade_authority_account = AccountSharedData::new(1, 0, &Pubkey::new_unique());
+            let transaction_accounts = vec![
+                (programdata_address, programdata_account),
+                (program_address, program_account),
+                (*buffer_address, buffer_account),
+                (spill_address, spill_account),
+                (sysvar::rent::id(), rent_account),
+                (sysvar::clock::id(), clock_account),
+                (*upgrade_authority_address, upgrade_authority_account),
+            ];
+            let instruction_accounts = vec![
+                AccountMeta {
+                    pubkey: programdata_address,
+                    is_signer: false,
+                    is_writable: true,
+                },
+                AccountMeta {
+                    pubkey: program_address,
+                    is_signer: false,
+                    is_writable: true,
+                },
+                AccountMeta {
+                    pubkey: *buffer_address,
+                    is_signer: false,
+                    is_writable: true,
+                },
+                AccountMeta {
+                    pubkey: spill_address,
+                    is_signer: false,
+                    is_writable: true,
+                },
+                AccountMeta {
+                    pubkey: sysvar::rent::id(),
+                    is_signer: false,
+                    is_writable: false,
+                },
+                AccountMeta {
+                    pubkey: sysvar::clock::id(),
+                    is_signer: false,
+                    is_writable: false,
+                },
+                AccountMeta {
+                    pubkey: *upgrade_authority_address,
+                    is_signer: true,
+                    is_writable: false,
+                },
+            ];
+            (transaction_accounts, instruction_accounts)
+        }
+
+        fn test_process_upgrade_instruction(
+            transaction_accounts: Vec<(Pubkey, AccountSharedData)>,
+            instruction_accounts: Vec<AccountMeta>,
+            expected_result: Result<(), InstructionError>,
+        ) -> Vec<AccountSharedData> {
+            let instruction_data =
+                bincode::serialize(&UpgradeableLoaderInstruction::Upgrade).unwrap();
+            mock_process_instruction(
+                &bpf_loader_upgradeable::id(),
+                Vec::new(),
+                &instruction_data,
+                transaction_accounts,
+                instruction_accounts,
+                None,
+                None,
+                expected_result,
+                process_instruction_without_jit,
+            )
+        }
+
+        // Case: Success
+        let (transaction_accounts, instruction_accounts) = get_accounts(
+            &buffer_address,
+            &upgrade_authority_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        let accounts =
+            test_process_upgrade_instruction(transaction_accounts, instruction_accounts, Ok(()));
+        let min_programdata_balance = Rent::default().minimum_balance(
+            UpgradeableLoaderState::size_of_programdata(elf_orig.len().max(elf_new.len())),
+        );
+        assert_eq!(
+            min_programdata_balance,
+            accounts.first().unwrap().lamports()
+        );
+        assert_eq!(0, accounts.get(2).unwrap().lamports());
+        assert_eq!(1, accounts.get(3).unwrap().lamports());
+        let state: UpgradeableLoaderState = accounts.first().unwrap().state().unwrap();
+        assert_eq!(
+            state,
+            UpgradeableLoaderState::ProgramData {
+                slot: SLOT,
+                upgrade_authority_address: Some(upgrade_authority_address)
+            }
+        );
+        for (i, byte) in accounts
+            .first()
+            .unwrap()
+            .data()
+            .get(
+                UpgradeableLoaderState::size_of_programdata_metadata()
+                    ..UpgradeableLoaderState::size_of_programdata(elf_new.len()),
+            )
+            .unwrap()
+            .iter()
+            .enumerate()
+        {
+            assert_eq!(*elf_new.get(i).unwrap(), *byte);
+        }
+
+        // Case: not upgradable
+        let (mut transaction_accounts, instruction_accounts) = get_accounts(
+            &buffer_address,
+            &upgrade_authority_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        transaction_accounts
+            .get_mut(0)
+            .unwrap()
+            .1
+            .set_state(&UpgradeableLoaderState::ProgramData {
+                slot: SLOT,
+                upgrade_authority_address: None,
+            })
+            .unwrap();
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::Immutable),
+        );
+
+        // Case: wrong authority
+        let (mut transaction_accounts, mut instruction_accounts) = get_accounts(
+            &buffer_address,
+            &upgrade_authority_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        let invalid_upgrade_authority_address = Pubkey::new_unique();
+        transaction_accounts.get_mut(6).unwrap().0 = invalid_upgrade_authority_address;
+        instruction_accounts.get_mut(6).unwrap().pubkey = invalid_upgrade_authority_address;
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::IncorrectAuthority),
+        );
+
+        // Case: authority did not sign
+        let (transaction_accounts, mut instruction_accounts) = get_accounts(
+            &buffer_address,
+            &upgrade_authority_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        instruction_accounts.get_mut(6).unwrap().is_signer = false;
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::MissingRequiredSignature),
+        );
+
+        // Case: Program account not executable
+        let (mut transaction_accounts, instruction_accounts) = get_accounts(
+            &buffer_address,
+            &upgrade_authority_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        transaction_accounts
+            .get_mut(1)
+            .unwrap()
+            .1
+            .set_executable(false);
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::AccountNotExecutable),
+        );
+
+        // Case: Program account now owned by loader
+        let (mut transaction_accounts, instruction_accounts) = get_accounts(
+            &buffer_address,
+            &upgrade_authority_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        transaction_accounts
+            .get_mut(1)
+            .unwrap()
+            .1
+            .set_owner(Pubkey::new_unique());
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::IncorrectProgramId),
+        );
+
+        // Case: Program account not writable
+        let (transaction_accounts, mut instruction_accounts) = get_accounts(
+            &buffer_address,
+            &upgrade_authority_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        instruction_accounts.get_mut(1).unwrap().is_writable = false;
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::InvalidArgument),
+        );
+
+        // Case: Program account not initialized
+        let (mut transaction_accounts, instruction_accounts) = get_accounts(
+            &buffer_address,
+            &upgrade_authority_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        transaction_accounts
+            .get_mut(1)
+            .unwrap()
+            .1
+            .set_state(&UpgradeableLoaderState::Uninitialized)
+            .unwrap();
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::InvalidAccountData),
+        );
+
+        // Case: Program ProgramData account mismatch
+        let (mut transaction_accounts, mut instruction_accounts) = get_accounts(
+            &buffer_address,
+            &upgrade_authority_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        let invalid_programdata_address = Pubkey::new_unique();
+        transaction_accounts.get_mut(0).unwrap().0 = invalid_programdata_address;
+        instruction_accounts.get_mut(0).unwrap().pubkey = invalid_programdata_address;
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::InvalidArgument),
+        );
+
+        // Case: Buffer account not initialized
+        let (mut transaction_accounts, instruction_accounts) = get_accounts(
+            &buffer_address,
+            &upgrade_authority_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        transaction_accounts
+            .get_mut(2)
+            .unwrap()
+            .1
+            .set_state(&UpgradeableLoaderState::Uninitialized)
+            .unwrap();
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::InvalidArgument),
+        );
+
+        // Case: Buffer account too big
+        let (mut transaction_accounts, instruction_accounts) = get_accounts(
+            &buffer_address,
+            &upgrade_authority_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        transaction_accounts.get_mut(2).unwrap().1 = AccountSharedData::new(
+            1,
+            UpgradeableLoaderState::size_of_buffer(
+                elf_orig.len().max(elf_new.len()).saturating_add(1),
+            ),
+            &bpf_loader_upgradeable::id(),
+        );
+        transaction_accounts
+            .get_mut(2)
+            .unwrap()
+            .1
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: Some(upgrade_authority_address),
+            })
+            .unwrap();
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::AccountDataTooSmall),
+        );
+
+        // Test small buffer account
+        let (mut transaction_accounts, instruction_accounts) = get_accounts(
+            &buffer_address,
+            &upgrade_authority_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        transaction_accounts
+            .get_mut(2)
+            .unwrap()
+            .1
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: Some(upgrade_authority_address),
+            })
+            .unwrap();
+        truncate_data(&mut transaction_accounts.get_mut(2).unwrap().1, 5);
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::InvalidAccountData),
+        );
+
+        // Case: Mismatched buffer and program authority
+        let (transaction_accounts, instruction_accounts) = get_accounts(
+            &buffer_address,
+            &buffer_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::IncorrectAuthority),
+        );
+
+        // Case: None buffer authority
+        let (mut transaction_accounts, instruction_accounts) = get_accounts(
+            &buffer_address,
+            &buffer_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        transaction_accounts
+            .get_mut(2)
+            .unwrap()
+            .1
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: None,
+            })
+            .unwrap();
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::IncorrectAuthority),
+        );
+
+        // Case: None buffer and program authority
+        let (mut transaction_accounts, instruction_accounts) = get_accounts(
+            &buffer_address,
+            &buffer_address,
+            &upgrade_authority_address,
+            &elf_orig,
+            &elf_new,
+        );
+        transaction_accounts
+            .get_mut(0)
+            .unwrap()
+            .1
+            .set_state(&UpgradeableLoaderState::ProgramData {
+                slot: SLOT,
+                upgrade_authority_address: None,
+            })
+            .unwrap();
+        transaction_accounts
+            .get_mut(2)
+            .unwrap()
+            .1
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: None,
+            })
+            .unwrap();
+        test_process_upgrade_instruction(
+            transaction_accounts,
+            instruction_accounts,
+            Err(InstructionError::IncorrectAuthority),
+        );
+    }
+
+    #[test]
+    fn test_set_upgrade_authority() {
+        let instruction = bincode::serialize(&UpgradeableLoaderInstruction::SetAuthority).unwrap();
+        let loader_id = bpf_loader_upgradeable::id();
+        let slot = 0;
+        let upgrade_authority_address = Pubkey::new_unique();
+        let upgrade_authority_account = AccountSharedData::new(1, 0, &Pubkey::new_unique());
+        let new_upgrade_authority_address = Pubkey::new_unique();
+        let new_upgrade_authority_account = AccountSharedData::new(1, 0, &Pubkey::new_unique());
+        let program_address = Pubkey::new_unique();
+        let (programdata_address, _) = Pubkey::find_program_address(
+            &[program_address.as_ref()],
+            &bpf_loader_upgradeable::id(),
+        );
+        let mut programdata_account = AccountSharedData::new(
+            1,
+            UpgradeableLoaderState::size_of_programdata(0),
+            &bpf_loader_upgradeable::id(),
+        );
+        programdata_account
+            .set_state(&UpgradeableLoaderState::ProgramData {
+                slot,
+                upgrade_authority_address: Some(upgrade_authority_address),
+            })
+            .unwrap();
+        let programdata_meta = AccountMeta {
+            pubkey: programdata_address,
+            is_signer: false,
+            is_writable: false,
+        };
+        let upgrade_authority_meta = AccountMeta {
+            pubkey: upgrade_authority_address,
+            is_signer: true,
+            is_writable: false,
+        };
+        let new_upgrade_authority_meta = AccountMeta {
+            pubkey: new_upgrade_authority_address,
+            is_signer: false,
+            is_writable: false,
+        };
+
+        // Case: Set to new authority
+        let accounts = test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![
+                (programdata_address, programdata_account.clone()),
+                (upgrade_authority_address, upgrade_authority_account.clone()),
+                (
+                    new_upgrade_authority_address,
+                    new_upgrade_authority_account.clone(),
+                ),
+            ],
+            vec![
+                programdata_meta.clone(),
+                upgrade_authority_meta.clone(),
+                new_upgrade_authority_meta.clone(),
+            ],
+            Ok(()),
+        );
+        let state: UpgradeableLoaderState = accounts.first().unwrap().state().unwrap();
+        assert_eq!(
+            state,
+            UpgradeableLoaderState::ProgramData {
+                slot,
+                upgrade_authority_address: Some(new_upgrade_authority_address),
+            }
+        );
+
+        // Case: Not upgradeable
+        let accounts = test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![
+                (programdata_address, programdata_account.clone()),
+                (upgrade_authority_address, upgrade_authority_account.clone()),
+            ],
+            vec![programdata_meta.clone(), upgrade_authority_meta.clone()],
+            Ok(()),
+        );
+        let state: UpgradeableLoaderState = accounts.first().unwrap().state().unwrap();
+        assert_eq!(
+            state,
+            UpgradeableLoaderState::ProgramData {
+                slot,
+                upgrade_authority_address: None,
+            }
+        );
+
+        // Case: Authority did not sign
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![
+                (programdata_address, programdata_account.clone()),
+                (upgrade_authority_address, upgrade_authority_account.clone()),
+            ],
+            vec![
+                programdata_meta.clone(),
+                AccountMeta {
+                    pubkey: upgrade_authority_address,
+                    is_signer: false,
+                    is_writable: false,
+                },
+            ],
+            Err(InstructionError::MissingRequiredSignature),
+        );
+
+        // Case: wrong authority
+        let invalid_upgrade_authority_address = Pubkey::new_unique();
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![
+                (programdata_address, programdata_account.clone()),
+                (
+                    invalid_upgrade_authority_address,
+                    upgrade_authority_account.clone(),
+                ),
+                (new_upgrade_authority_address, new_upgrade_authority_account),
+            ],
+            vec![
+                programdata_meta.clone(),
+                AccountMeta {
+                    pubkey: invalid_upgrade_authority_address,
+                    is_signer: true,
+                    is_writable: false,
+                },
+                new_upgrade_authority_meta,
+            ],
+            Err(InstructionError::IncorrectAuthority),
+        );
+
+        // Case: No authority
+        programdata_account
+            .set_state(&UpgradeableLoaderState::ProgramData {
+                slot,
+                upgrade_authority_address: None,
+            })
+            .unwrap();
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![
+                (programdata_address, programdata_account.clone()),
+                (upgrade_authority_address, upgrade_authority_account.clone()),
+            ],
+            vec![programdata_meta.clone(), upgrade_authority_meta.clone()],
+            Err(InstructionError::Immutable),
+        );
+
+        // Case: Not a ProgramData account
+        programdata_account
+            .set_state(&UpgradeableLoaderState::Program {
+                programdata_address: Pubkey::new_unique(),
+            })
+            .unwrap();
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![
+                (programdata_address, programdata_account.clone()),
+                (upgrade_authority_address, upgrade_authority_account),
+            ],
+            vec![programdata_meta, upgrade_authority_meta],
+            Err(InstructionError::InvalidArgument),
+        );
+    }
+
+    #[test]
+    fn test_set_buffer_authority() {
+        let instruction = bincode::serialize(&UpgradeableLoaderInstruction::SetAuthority).unwrap();
+        let loader_id = bpf_loader_upgradeable::id();
+        let invalid_authority_address = Pubkey::new_unique();
+        let authority_address = Pubkey::new_unique();
+        let authority_account = AccountSharedData::new(1, 0, &Pubkey::new_unique());
+        let new_authority_address = Pubkey::new_unique();
+        let new_authority_account = AccountSharedData::new(1, 0, &Pubkey::new_unique());
+        let buffer_address = Pubkey::new_unique();
+        let mut buffer_account =
+            AccountSharedData::new(1, UpgradeableLoaderState::size_of_buffer(0), &loader_id);
+        buffer_account
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: Some(authority_address),
+            })
+            .unwrap();
+        let mut transaction_accounts = vec![
+            (buffer_address, buffer_account.clone()),
+            (authority_address, authority_account.clone()),
+            (new_authority_address, new_authority_account.clone()),
+        ];
+        let buffer_meta = AccountMeta {
+            pubkey: buffer_address,
+            is_signer: false,
+            is_writable: false,
+        };
+        let authority_meta = AccountMeta {
+            pubkey: authority_address,
+            is_signer: true,
+            is_writable: false,
+        };
+        let new_authority_meta = AccountMeta {
+            pubkey: new_authority_address,
+            is_signer: false,
+            is_writable: false,
+        };
+
+        // Case: New authority required
+        let accounts = test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            transaction_accounts.clone(),
+            vec![buffer_meta.clone(), authority_meta.clone()],
+            Err(InstructionError::IncorrectAuthority),
+        );
+        let state: UpgradeableLoaderState = accounts.first().unwrap().state().unwrap();
+        assert_eq!(
+            state,
+            UpgradeableLoaderState::Buffer {
+                authority_address: Some(authority_address),
+            }
+        );
+
+        // Case: Set to new authority
+        buffer_account
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: Some(authority_address),
+            })
+            .unwrap();
+        let accounts = test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            transaction_accounts.clone(),
+            vec![
+                buffer_meta.clone(),
+                authority_meta.clone(),
+                new_authority_meta.clone(),
+            ],
+            Ok(()),
+        );
+        let state: UpgradeableLoaderState = accounts.first().unwrap().state().unwrap();
+        assert_eq!(
+            state,
+            UpgradeableLoaderState::Buffer {
+                authority_address: Some(new_authority_address),
+            }
+        );
+
+        // Case: Authority did not sign
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            transaction_accounts.clone(),
+            vec![
+                buffer_meta.clone(),
+                AccountMeta {
+                    pubkey: authority_address,
+                    is_signer: false,
+                    is_writable: false,
+                },
+                new_authority_meta.clone(),
+            ],
+            Err(InstructionError::MissingRequiredSignature),
+        );
+
+        // Case: wrong authority
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![
+                (buffer_address, buffer_account.clone()),
+                (invalid_authority_address, authority_account),
+                (new_authority_address, new_authority_account),
+            ],
+            vec![
+                buffer_meta.clone(),
+                AccountMeta {
+                    pubkey: invalid_authority_address,
+                    is_signer: true,
+                    is_writable: false,
+                },
+                new_authority_meta.clone(),
+            ],
+            Err(InstructionError::IncorrectAuthority),
+        );
+
+        // Case: No authority
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            transaction_accounts.clone(),
+            vec![buffer_meta.clone(), authority_meta.clone()],
+            Err(InstructionError::IncorrectAuthority),
+        );
+
+        // Case: Set to no authority
+        transaction_accounts
+            .get_mut(0)
+            .unwrap()
+            .1
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: None,
+            })
+            .unwrap();
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            transaction_accounts.clone(),
+            vec![
+                buffer_meta.clone(),
+                authority_meta.clone(),
+                new_authority_meta.clone(),
+            ],
+            Err(InstructionError::Immutable),
+        );
+
+        // Case: Not a Buffer account
+        transaction_accounts
+            .get_mut(0)
+            .unwrap()
+            .1
+            .set_state(&UpgradeableLoaderState::Program {
+                programdata_address: Pubkey::new_unique(),
+            })
+            .unwrap();
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            transaction_accounts.clone(),
+            vec![buffer_meta, authority_meta, new_authority_meta],
+            Err(InstructionError::InvalidArgument),
+        );
+    }
+
+    #[test]
+    fn test_close() {
+        let instruction = bincode::serialize(&UpgradeableLoaderInstruction::Close).unwrap();
+        let loader_id = bpf_loader_upgradeable::id();
+        let invalid_authority_address = Pubkey::new_unique();
+        let authority_address = Pubkey::new_unique();
+        let authority_account = AccountSharedData::new(1, 0, &Pubkey::new_unique());
+        let recipient_address = Pubkey::new_unique();
+        let recipient_account = AccountSharedData::new(1, 0, &Pubkey::new_unique());
+        let buffer_address = Pubkey::new_unique();
+        let mut buffer_account =
+            AccountSharedData::new(1, UpgradeableLoaderState::size_of_buffer(0), &loader_id);
+        buffer_account
+            .set_state(&UpgradeableLoaderState::Buffer {
+                authority_address: Some(authority_address),
+            })
+            .unwrap();
+        let uninitialized_address = Pubkey::new_unique();
+        let mut uninitialized_account = AccountSharedData::new(
+            1,
+            UpgradeableLoaderState::size_of_programdata(0),
+            &loader_id,
+        );
+        uninitialized_account
+            .set_state(&UpgradeableLoaderState::Uninitialized)
+            .unwrap();
+        let programdata_address = Pubkey::new_unique();
+        let mut programdata_account = AccountSharedData::new(
+            1,
+            UpgradeableLoaderState::size_of_programdata(0),
+            &loader_id,
+        );
+        programdata_account
+            .set_state(&UpgradeableLoaderState::ProgramData {
+                slot: 0,
+                upgrade_authority_address: Some(authority_address),
+            })
+            .unwrap();
+        let program_address = Pubkey::new_unique();
+        let mut program_account =
+            AccountSharedData::new(1, UpgradeableLoaderState::size_of_program(), &loader_id);
+        program_account.set_executable(true);
+        program_account
+            .set_state(&UpgradeableLoaderState::Program {
+                programdata_address,
+            })
+            .unwrap();
+        let transaction_accounts = vec![
+            (buffer_address, buffer_account.clone()),
+            (recipient_address, recipient_account.clone()),
+            (authority_address, authority_account.clone()),
+        ];
+        let buffer_meta = AccountMeta {
+            pubkey: buffer_address,
+            is_signer: false,
+            is_writable: false,
+        };
+        let recipient_meta = AccountMeta {
+            pubkey: recipient_address,
+            is_signer: false,
+            is_writable: false,
+        };
+        let authority_meta = AccountMeta {
+            pubkey: authority_address,
+            is_signer: true,
+            is_writable: false,
+        };
+
+        // Case: close a buffer account
+        let accounts = test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            transaction_accounts,
+            vec![
+                buffer_meta.clone(),
+                recipient_meta.clone(),
+                authority_meta.clone(),
+            ],
+            Ok(()),
+        );
+        assert_eq!(0, accounts.first().unwrap().lamports());
+        assert_eq!(2, accounts.get(1).unwrap().lamports());
+        let state: UpgradeableLoaderState = accounts.first().unwrap().state().unwrap();
+        assert_eq!(state, UpgradeableLoaderState::Uninitialized);
+
+        // Case: close with wrong authority
+        test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![
+                (buffer_address, buffer_account.clone()),
+                (recipient_address, recipient_account.clone()),
+                (invalid_authority_address, authority_account.clone()),
+            ],
+            vec![
+                buffer_meta,
+                recipient_meta.clone(),
+                AccountMeta {
+                    pubkey: invalid_authority_address,
+                    is_signer: true,
+                    is_writable: false,
+                },
+            ],
+            Err(InstructionError::IncorrectAuthority),
+        );
+
+        // Case: close an uninitialized account
+        let accounts = test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![
+                (uninitialized_address, uninitialized_account.clone()),
+                (recipient_address, recipient_account.clone()),
+                (invalid_authority_address, authority_account.clone()),
+            ],
+            vec![
+                AccountMeta {
+                    pubkey: uninitialized_address,
+                    is_signer: false,
+                    is_writable: false,
+                },
+                recipient_meta.clone(),
+                authority_meta.clone(),
+            ],
+            Ok(()),
+        );
+        assert_eq!(0, accounts.first().unwrap().lamports());
+        assert_eq!(2, accounts.get(1).unwrap().lamports());
+        let state: UpgradeableLoaderState = accounts.first().unwrap().state().unwrap();
+        assert_eq!(state, UpgradeableLoaderState::Uninitialized);
+
+        // Case: close a program account
+        let accounts = test_process_instruction(
+            &loader_id,
+            &[],
+            &instruction,
+            vec![
+                (programdata_address, programdata_account.clone()),
+                (recipient_address, recipient_account),
+                (authority_address, authority_account),
+                (program_address, program_account.clone()),
+            ],
+            vec![
+                AccountMeta {
+                    pubkey: programdata_address,
+                    is_signer: false,
+                    is_writable: false,
+                },
+                recipient_meta,
+                authority_meta,
+                AccountMeta {
+                    pubkey: program_address,
+                    is_signer: false,
+                    is_writable: true,
+                },
+            ],
+            Ok(()),
+        );
+        assert_eq!(0, accounts.first().unwrap().lamports());
+        assert_eq!(2, accounts.get(1).unwrap().lamports());
+        let state: UpgradeableLoaderState = accounts.first().unwrap().state().unwrap();
+        assert_eq!(state, UpgradeableLoaderState::Uninitialized);
+
+        // Try to invoke closed account
+        test_process_instruction(
+            &program_address,
+            &[0, 1],
+            &[],
+            vec![
+                (programdata_address, programdata_account.clone()),
+                (program_address, program_account.clone()),
+            ],
+            Vec::new(),
+            Err(InstructionError::InvalidAccountData),
+        );
+    }
+}
